imagedownloader/requirements.txt
Beaker==1.5.4
Django==1.4.3
IPy==0.75
PyVirtualDisplay==0.1.1
PyYAML==3.10
South==0.7.6
decorator==3.4.0
distribute==0.6.27
django-model-utils==1.2.0
docutils==0.8.1
email
numpy
h5py==2.1.2
ipython==0.12
mglob==0.4
netCDF4==1.0.2
psycopg2==2.4.5
python-dateutil==1.5
pytz==2012j
scipy==0.10.1
splinter==0.4.9
